+++ tests/deprecated_api/test_remove_1-4.py
@@ -89,22 +89,6 @@ def test_v1_4_0_deprecated_trainer_device_distrib():
         assert trainer.use_horovod
 
 
-def test_v1_4_0_deprecated_trainer_phase():
-    """Test that Trainer attributes works fine."""
-    trainer = Trainer()
-
-    assert not trainer.training
-    assert not trainer.testing
-
-    trainer.training = True
-    assert trainer.training
-    assert not trainer.testing
-
-    trainer.testing = True
-    assert not trainer.training
-    assert trainer.testing
-
-
 def test_v1_4_0_deprecated_metrics():
     from pytorch_lightning.metrics.functional.classification import stat_scores_multiple_classes
     with pytest.deprecated_call(match='will be removed in v1.4'):